pixels: properly set width and CELL_PIXEL_GRAPHICS #1397
